+++ Glyssen/Dialogs/AddCharactersToGroupViewModel.cs
@@ -26,7 +26,7 @@ namespace Glyssen.Dialogs
 				m_cameoActor = cameoActor;
 				foreach (var kvp in charactersNotInGroup)
 				{
-					if ((kvp.Value / (double) Program.kKeyStrokesPerHour) < Program.kCameoCharacterEstimatedHoursLimit)
+					if ((kvp.Value / Program.kKeyStrokesPerHour) < Program.kCameoCharacterEstimatedHoursLimit)
 					{
 						var detail = m_characterDetailDictionary[kvp.Key];
 						if (cameoActor.Matches(detail))
@@ -90,7 +90,7 @@ namespace Glyssen.Dialogs
 		public double GetEstimatedHoursForCharacter(int filteredCharacterIndex)
 		{
 			return m_keyStrokesByCharacterId[ActiveList[filteredCharacterIndex].CharacterId] /
-				(double) Program.kKeyStrokesPerHour;
+				Program.kKeyStrokesPerHour;
 		}
 	}
 }
