+++ Glyssen/Program.cs
@@ -20,7 +20,7 @@ namespace Glyssen
 		public const string kProduct = "Glyssen";
 		public const string kApplicationId = "Glyssen";
 
-		public const int kKeyStrokesPerHour = 6000;
+		public const double kKeyStrokesPerHour = 6000;
 		public const double kCameoCharacterEstimatedHoursLimit = 0.2;
 
 		private const string kOldProductName = "Protoscript Generator";
